Create acceptance_strategy combined-init
Add logic like mentioned in report 2. Not quite sure if the time used in every call is correct, but followed the report and it makes sense like this. Added parentheses to make the logic more easy to understand, and also to make sure its read in the correct way. Also the acceptance_init and acceptance_combined is divided up in two different if statements, to make it a bit more easy to understand.